+++ src/caselawclient/responses/search_result.py
@@ -194,6 +194,13 @@ class SearchResult:
         """
 
         court_code = self._get_xpath_match_string("search:extracted/uk:court/text()")
+        jurisdiction_code = self._get_xpath_match_string(
+            "search:extracted/uk:jurisdiction/text()"
+        )
+
+        if jurisdiction_code:
+            court_code = "%s/%s" % (court_code, jurisdiction_code)
+
         try:
             court = courts.get_by_code(court_code)
         except CourtNotFoundException: